Uploaded the entire project done in scratch
Project done in scratch as a part of CS50 course on HarvardX.

My scratch project is a click-to-survive game where the main character is attacked by spawning ghosts. The game is started by clicking the space button. Player can kill ghosts by clicking on them before they reach the main character. Difficulty increases with time; more ghosts are spawned and they spawn faster. At the start of the game the player has 3 lives, when all 3 are lost the game is over. The game can be restarted by clicking space again.